+++ package.json
@@ -11,7 +11,6 @@
     "lint:prod": "eslint '**/*.ts' --max-warnings 0 && eslint '*.js' '{controllers,models,router,services,seeds,worker}/**/*.js' --no-eslintrc --config ./.eslintrc-javascript.js --max-warnings 0",
     "worker": "ts-node worker",
     "worker:dev": "ts-node-dev worker",
-    "add-cron": "ts-node worker/add-cron"
   },
   "nyc": {
     "check-coverage": true,
@@ -82,6 +81,7 @@
   "devDependencies": {
     "@types/bull": "^3.12.2",
     "@types/lodash": "^4.14.149",
+    "@types/mongoose": "^5.7.10",
     "@typescript-eslint/eslint-plugin": "^2.27.0",
     "@typescript-eslint/parser": "^2.27.0",
     "ava": "^1.3.1",
